Moved cli stuff back
